requirements.txt
# ── Core ────────────────────────────────────────
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
yfinance>=0.2.40
pydantic>=2.6.0
python-docx>=1.1.0
feedparser>=6.0.0

# ── LLM ─────────────────────────────────────────
google-generativeai>=0.5.0

# ── Dev / Test ──────────────────────────────────
pytest>=8.0.0
python-dotenv>=1.0.0

# ── JS / headless browser (SCMP search, JS-rendered pages) ─────
playwright>=1.42.0

# ── Web API (Render / frontend) ─────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
